+++ src/data/types/nucleotides/view_traits.rs
@@ -1,9 +1,9 @@
 use crate::{
-    data::views::{impl_len, impl_restrict_for_wrapper, impl_slice_for_wrapper, impl_views_for_wrapper},
+    data::views::{impl_len_for_views, impl_restrict_for_wrapper, impl_slice_for_wrapper, impl_views_for_wrapper},
     prelude::*,
 };
 
-impl_len!(Nucleotides, NucleotidesView, NucleotidesViewMut, 0);
+impl_len_for_views!(Nucleotides, NucleotidesView, NucleotidesViewMut, 0);
 impl_views_for_wrapper!(Nucleotides, NucleotidesView, NucleotidesViewMut);
 impl_slice_for_wrapper!(Nucleotides, NucleotidesView, NucleotidesViewMut);
 impl_restrict_for_wrapper!(Nucleotides, NucleotidesView, NucleotidesViewMut);