+++ CHANGELOG.md
@@ -8,7 +8,9 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 
 ### Added
 
+- Added `get_aligned_query`, similar to `get_aligned_seqs` but returning only the query sequence
 - Added `from_ciglets_unchecked` and `from_cigar_unchecked` to `AlignmentStates`
+- Introduced the `AlignmentAndSeqs` arbitrary wrapper (behind `fuzzing` feature gate)
 - Added `visit_params` methods to pHMMs as a diagnostic tool for inspecting the parameters that are visited along a given path (behind `dev-phmm` and `alignment-diagnostics` feature gates)
 - Added the ability to remove layers from a pHMM (behind `dev-phmm` and `alignment-diagnostics` feature gates)
 - Added random downsampling for sized and non-sized iterators using `MethodDSampler`, `method_l`, and `BernoulliSampler`
@@ -18,6 +20,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - `CigletIteratorChecked` now requires that adjacent ciglets have differing operations. This also impacts `TryFrom` impls and `is_valid`
 - An iterator of ciglets now collects into a `Result<AlignmentStates, CigarError>` instead of `AlignmentStates`
 - `sw_score_from_path` is now behind the `alignment-diagnostics` feature gate
+- Modified the behavior of `AlignmentArbitrary` (behind `fuzzing` feature gate)
 - Refactored the organization and visibility of pHMM structs/fields/modules (behind `dev-phmm` feature gate)
 - Modified Viterbi algorithm to use compact bit representation instead of array (behind `dev-phmm` feature gate)